sol-idx/lambda/short_requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
idna==3.3
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.19.2
requests==2.27.1
urllib3==1.26.8
